Refactoring (#16)
* buffers all cleaned up

* done

* Set up ImGui (#15)

* feat: basic initialize/free functions for gui

* feat: add material loading onto the CPU

loadGLTF now loads materials and textures into Scene.materials and Scene.textures

* feat: get it to work with manual files

* feat: remove imgui from vcpkg, add to separate `third_party` folder

* fix: duplicate variable inits

* format: destructor_stack

* feat: switch to destructor stack for imgui's descriptor pool

* feat: move gui-related code out of main.cpp, into specialized functions

* feat: FPS counter

* format: all files

* feat: fix merge conflicts

---------


* capitalize engine

---------

+++ src/scene/scene.cpp
@@ -324,7 +324,7 @@ bool load_gltf( std::string filepath,
                         "with pos count" );
                 }
 
-                new_prim.vertex_offset = out_global_vertices.size();
+                new_prim.vertex_offset = static_cast<int>( out_global_vertices.size() );
 
                 for ( size_t i = 0; i < pos.size(); ++i ) {
                     geometry::Vertex new_vertex;
@@ -359,7 +359,7 @@ bool load_gltf( std::string filepath,
 
                     tinygltf::BufferView buffer_view =
                         model.bufferViews[static_cast<size_t>( buffer_view_id )];
-                    new_prim.ind_offset = out_global_indices.size();
+                    new_prim.ind_offset = static_cast<int>( out_global_indices.size() );
                     new_prim.ind_count = accessor.count;
                     size_t byte_offset = buffer_view.byteOffset;
                     size_t byte_length = buffer_view.byteLength;
@@ -391,7 +391,7 @@ bool load_gltf( std::string filepath,
                           // triangles
                     new_prim.is_indexed = false;
                     for ( size_t i = 0; i < pos.size(); i++ ) {
-                        out_global_indices.push_back( i );
+                        out_global_indices.push_back( static_cast<unsigned int>( i ) );
                     }
                 }
